Replace unique_signals json macro with serde map
Replaced the code that stored unique signals to use a serde_json::Map because using the json macro resulted in some annoying behavior having to call `.as_object()` to do anything with it. Using a map directly removes the need while still allowing serialization at the end. This should probably slightly increase performance too, but no difference was noticed when I made the change on my end and built for release